requirements/requirements.txt
asgiref==3.5.2
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
boto3==1.24.9
botocore==1.27.9
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
Django==3.2.13
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.9.0
parsel==1.6.0
priority==1.3.0
Protego==0.2.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
queuelib==1.6.2
requests==2.28.0
requests-pkcs12==1.14
requests-toolbelt==0.9.1
s3transfer==0.6.0
Scrapy==2.5.1
scrapy-splash==0.8.0
sentry-sdk==0.20.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
Werkzeug==1.0.1
whitenoise==6.2.0
zope.interface==5.4.0
crochet==2.1.1
pycurl
